refactor getMissionsForRanking
