+++ src/app/seasons/[slug]/ranking/mission/page.tsx
@@ -1,3 +1,4 @@
+import { getMissionsForRanking } from "@/features/missions/services/missions";
 import { CurrentUserCardMission } from "@/features/ranking/components/current-user-card-mission";
 import { MissionSelect } from "@/features/ranking/components/mission-select";
 import { RankingMission } from "@/features/ranking/components/ranking-mission";
@@ -9,7 +10,6 @@ import {
 } from "@/features/ranking/services/get-missions-ranking";
 import { getUser } from "@/features/user-profile/services/profile";
 import { getSeasonBySlug } from "@/lib/services/seasons";
-import { createClient } from "@/lib/supabase/client";
 import type { Metadata } from "next";
 import { notFound } from "next/navigation";
 
@@ -50,39 +50,13 @@ export default async function SeasonMissionRankingPage({
     notFound();
   }
 
-  const supabase = createClient();
-
   // ユーザー情報取得
   const user = await getUser();
 
-  // ミッション一覧をカテゴリ情報付きで取得（max_achievement_countがnullのもののみ）
-  const { data: missions, error: missionsError } = await supabase
-    .from("missions")
-    .select(`
-      *,
-      mission_category_link(
-        mission_category(
-          id,
-          category_title,
-          sort_no
-        )
-      )
-    `)
-    .is("max_achievement_count", null)
-    .order("is_featured", { ascending: false }) // is_featuredがtrueのものを先頭に
-    .order("difficulty", { ascending: true }); // その後、難易度の昇順でソート
+  // ミッション一覧を取得
+  const missions = await getMissionsForRanking();
 
-  // エラーハンドリング
-  if (missionsError) {
-    console.error("ミッション取得エラー:", missionsError);
-    return (
-      <div className="p-4 text-red-600">
-        ミッションの取得中にエラーが発生しました。
-      </div>
-    );
-  }
-
-  if (!missions || missions.length === 0) {
+  if (missions.length === 0) {
     return (
       <div className="p-4 text-gray-600">
         現在利用可能なミッションがありません。
